+++ src/Core/Types/Types/ResolveAbstractType.cs
